pyproject.toml
[project]
name = "timeline"
version = "0.1.0"
description = "Multi-tenant event sourcing system with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core Framework
    "fastapi[standard]>=0.128.0",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.21",
    # Database & ORM
    "sqlalchemy>=2.0.45",
    "asyncpg>=0.31.0",
    "greenlet>=3.0.0",
    # Caching
    "redis>=7.1.0",
    # Authentication & Security
    "authlib>=1.6.6",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "python-jose>=3.5.0",
    # Email Integration
    "aioimaplib>=2.0.1",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.45.0",
    "msal>=1.34.0",
    # Storage
    "aiofiles>=25.1.0",
    "boto3>=1.42.19",
    # Background Tasks
    "celery>=5.6.1",
    # Utilities
    "cuid2>=2.0.1",
    "jsonschema>=4.25.1",
    "slowapi>=0.1.9",
    # Observability
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-jaeger>=1.11.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-logging>=0.60b1",
    "opentelemetry-instrumentation-redis>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "freezegun>=1.5.1",
    "aioboto3>=13.0.0",
    "httpx>=0.27.0",

    # Code Quality
    "black>=25.12.0",
    "isort>=7.0.0",
    "autoflake>=2.3.1",
    "flake8>=7.0.0",
    "mypy>=1.11.0",

    # Database Migrations
    "alembic>=1.14.0",
]
